build: Add dependent repo fetch script [skip ci]
Add a Python program script that clones and builds other
repositories that this repository depends on.

The script is accompanied by a JSON database that contains
the repository information, "known-good" revision information,
and other build parameters for each dependent repository.

The dependent repositories are checked out to the "known-good"
revisions to ensure that they are compatible with the checked-out
revision of this repository.

See the BUILD.md file, the update_deps.py --help output, and
the documentation found inside of the update_deps.py file for
more information.

Fixes #161
Fixes #169